+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## CHANGELOG
 
-1. better images and video styles (using object-fit & max-width,height), yes even the youtube ones.
+1. new styles for `shift + /` keymapping help menu.